backend/requirements.txt
# Drone Simulation Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Socket.IO for real-time communication
python-socketio>=5.10.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Scientific computing
numpy>=1.26.0
scipy>=1.12.0

# Gesture recognition (Phase 5)
mediapipe>=0.10.9
opencv-python>=4.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
python-dotenv>=1.0.0
